refactor: extract article search SQL and helpers from route
The route mixed a 193-line ranked full-text SQL query with param
parsing, hit validation, and multi-collection hydration.

Move all three into src/features/articles/public/search/, alongside the
existing scope/indexItem modules the hydration already depends on. The
SQL constant is now a named module rather than a wall of text the reader
must scroll past to find the request handling.

Pure extraction — no behavior change. The SQL string is byte-identical
(verified programmatically); helper bodies are unchanged.

route.ts: 339 -> 65 lines.

Verified: 344/344 tests pass; tsc error signature identical to baseline;
eslint clean.
